Move classes into packages, and fixed a minor bug of tasks not being added to storage.txt
