+++ .gitignore
@@ -15,4 +15,4 @@ bin/
 
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
-src/main/java/chatbot/storage/list.txt
+src/main/java/rocket/storage/storage.txt
