--- src/main/java/chatbot/TaskType.java
@@ -1,5 +1,0 @@
-package chatbot;
-
-public enum TaskType {
-    TODO, DEADLINE, EVENT
-}
