--- src/main/java/chatbot/command/ExitCommand.java
@@ -1,16 +1,0 @@
-package chatbot.command;
-
-import chatbot.TaskList;
-import chatbot.Ui;
-import chatbot.chatbot.storage.Storage;
-
-public class ExitCommand extends Command {
-    public ExitCommand() {
-        super(true);
-    }
-
-    @Override
-    public void execute(TaskList list, Ui ui, Storage storage) {
-        ui.bye();
-    }
-}
