--- src/main/java/chatbot/command/InvalidFormatCommand.java
@@ -1,15 +1,0 @@
-package chatbot.command;
-
-import chatbot.TaskList;
-import chatbot.Ui;
-import chatbot.chatbot.storage.Storage;
-
-public class InvalidFormatCommand extends Command {
-    public InvalidFormatCommand() {
-        super(false);
-    }
-    @Override
-    public void execute(TaskList list, Ui ui, Storage storage) {
-        ui.readInvalidFormat();
-    }
-}
