--- src/main/java/chatbot/common/Utils.java
@@ -1,12 +1,0 @@
-package chatbot.common;
-
-public class Utils {
-    public static boolean isInteger(String x) {
-        try {
-            int intVal = Integer.parseInt(x);
-            return true;
-        } catch (NumberFormatException e) {
-            return false;
-        }
-    }
-}
