--- src/main/java/chatbot/exception/EmptyTaskNameException.java
@@ -1,7 +1,0 @@
-package chatbot.exception;
-
-public class EmptyTaskNameException extends RuntimeException {
-    public EmptyTaskNameException(String message) {
-        super(message);
-    }
-}
