--- src/main/java/chatbot/exception/RocketException.java
@@ -1,7 +1,0 @@
-package chatbot.exception;
-
-public class RocketException extends Exception {
-    public RocketException(String message) {
-        super(message);
-    }
-}
