--- src/main/java/chatbot/exception/RocketRuntimeException.java
@@ -1,7 +1,0 @@
-package chatbot.exception;
-
-public class RocketRuntimeException extends RuntimeException{
-    public RocketRuntimeException(String message) {
-        super(message);
-    }
-}
