+++ src/main/java/rocket/command/Command.java
@@ -1,0 +1,19 @@
+package rocket.command;
+
+import rocket.storage.Storage;
+import rocket.task.TaskList;
+import rocket.ui.Ui;
+
+public abstract class Command {
+    private final boolean isExit; // Checks the command is an exit command
+
+    public Command(boolean isExit) {
+        this.isExit = isExit;
+    }
+
+    public boolean isExit() {
+        return this.isExit;
+    }
+
+    public abstract void execute(TaskList list, Ui ui, Storage storage);
+}
