+++ src/main/java/rocket/command/ExitCommand.java
@@ -1,0 +1,16 @@
+package rocket.command;
+
+import rocket.task.TaskList;
+import rocket.ui.Ui;
+import rocket.storage.Storage;
+
+public class ExitCommand extends Command {
+    public ExitCommand() {
+        super(true);
+    }
+
+    @Override
+    public void execute(TaskList list, Ui ui, Storage storage) {
+        ui.bye();
+    }
+}
