+++ src/main/java/rocket/command/ListCommand.java
@@ -1,0 +1,20 @@
+package rocket.command;
+
+import rocket.task.TaskList;
+import rocket.ui.Ui;
+import rocket.storage.Storage;
+
+public class ListCommand extends Command {
+    public ListCommand() {
+        super(false);
+    }
+
+    @Override
+    public void execute(TaskList list, Ui ui, Storage storage) {
+        if (list.getSize() == 0) {
+            ui.readEmptyList();
+        } else {
+            ui.readListItems(list.getTasks());
+        }
+    }
+}
