+++ src/main/java/rocket/common/Utils.java
@@ -1,0 +1,12 @@
+package rocket.common;
+
+public class Utils {
+    public static boolean isInteger(String x) {
+        try {
+            int intVal = Integer.parseInt(x);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
+}
