+++ src/main/java/rocket/exception/EmptyTaskNameException.java
@@ -1,0 +1,7 @@
+package rocket.exception;
+
+public class EmptyTaskNameException extends RuntimeException {
+    public EmptyTaskNameException(String message) {
+        super(message);
+    }
+}
