+++ src/main/java/rocket/exception/RocketException.java
@@ -1,0 +1,7 @@
+package rocket.exception;
+
+public class RocketException extends Exception {
+    public RocketException(String message) {
+        super(message);
+    }
+}
