+++ src/main/java/rocket/exception/RocketRuntimeException.java
@@ -1,0 +1,7 @@
+package rocket.exception;
+
+public class RocketRuntimeException extends RuntimeException{
+    public RocketRuntimeException(String message) {
+        super(message);
+    }
+}
