+++ src/main/java/rocket/task/TaskType.java
@@ -1,0 +1,5 @@
+package rocket.task;
+
+public enum TaskType {
+    TODO, DEADLINE, EVENT
+}
